Fix integer overflow in MemoryPressureLevelReporter
Use base::saturated_cast<int>() instead of implicit int64_t to int
conversion in ReportHistogram() to prevent integer overflow when
duration exceeds INT_MAX.

Bug: 429856461
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6701163
Commit-Queue: Ho Cheung <hocheung@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1542557}

+++ components/memory_pressure/memory_pressure_level_reporter.cc
@@ -8,6 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/functional/bind.h"
 #include "base/memory/memory_pressure_listener.h"
 #include "base/metrics/histogram_functions.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/strings/strcat.h"
 #include "base/task/sequenced_task_runner.h"
 #include "base/time/time.h"
@@ -103,7 +104,8 @@ void MemoryPressureLevelReporter::ReportHistogram(base::TimeTicks now) {
         "Memory.PressureLevel2", 1, base::MemoryPressureLevel::kMaxValue + 1,
         base::MemoryPressureLevel::kMaxValue + 2,
         base::HistogramBase::kUmaTargetedHistogramFlag)
-        ->AddCount(current_pressure_level_, duration_s);
+        ->AddCount(current_pressure_level_,
+                   base::saturated_cast<int>(duration_s));
   }
 }
